Update README.md
I found it a little hard to read the generated html at a glance in the old readme. Added some syntax highlighting to try and improve legibility. I assume that technically the generated HTML is on one line but I figured adding the indentation made it easier for someone to grok what would be generated.